+++ cmd/upgrade.go
@@ -9,13 +9,13 @@ var upgradeCmd = &cobra.Command{
 	Use:   "upgrade",
 	Short: "Upgrade autorestic and restic",
 	Run: func(cmd *cobra.Command, args []string) {
-		noRestic, _ := cmd.Flags().GetBool("no-restic")
-		err := bins.Upgrade(!noRestic)
+		restic, _ := cmd.Flags().GetBool("restic")
+		err := bins.Upgrade(restic)
 		CheckErr(err)
 	},
 }
 
 func init() {
 	rootCmd.AddCommand(upgradeCmd)
-	upgradeCmd.Flags().Bool("no-restic", false, "also update restic")
+	upgradeCmd.Flags().Bool("restic", true, "also update restic")
 }